requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.2
uuid==1.30
pytest==7.4.3
pytest-asyncio==0.23.2
httpx==0.25.2
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0 